+++ tests/test_window.rs
@@ -455,6 +455,40 @@ fn window_dense_rank() {
 }
 
 #[test]
+fn window_rank() {
+    let wrk = Workdir::new("window_rank");
+    wrk.create(
+        "numbers.csv",
+        vec![
+            svec!["n"],
+            svec!["20"],
+            svec!["10"],
+            svec!["30"],
+            svec!["10"],
+            svec!["20"],
+            svec!["20"],
+            svec!["20"],
+        ],
+    );
+    let mut cmd = wrk.command("window");
+    cmd.arg("rank(n) as rank").arg("numbers.csv");
+
+    let got: Vec<Vec<String>> = wrk.read_stdout(&mut cmd);
+    let expected = vec![
+        svec!["n", "rank"],
+        svec!["20", "3"],
+        svec!["10", "1"],
+        svec!["30", "7"],
+        svec!["10", "2"],
+        svec!["20", "4"],
+        svec!["20", "5"],
+        svec!["20", "6"],
+    ];
+
+    assert_eq!(got, expected);
+}
+
+#[test]
 fn window_generic_agg() {
     let wrk = Workdir::new("window_generic_agg");
     wrk.create(